Relax valgrind suppression, enable 24.04 in CI (#11)
It appears that the call stack related to the perceived memory leak
related to stack allocation in threads has changed slightly in Ubuntu
24.04. This change relaxes that suppression slightly and enables 24.04
builds by default.

Independently, this suppression needs to be relaxed in the same way to
be applicable to std::thread instances used in testing, for which the
same perceived leak manifests.