Introduce CMake find modules for dependencies
While this approach maintains pkg-config as the preferred mechanism for
discovering dependencies, it allows us to support the older config
mechanism shipped with many of these dependencies on older distributions
like CentOS 7.

An additional part of this change is the switch to CMake targets for
dependencies, which significantly reduces boilerplate in the main
CMakeLists.txt.