Attempt graceful shutdown on SIGTERM and SIGINT (#15)
The existing behavior is to terminate the program immediately when
receiving SIGTERM or SIGINT. While this will *probably* be okay, it
would be better if we attempt to wind down the process gracefully.

To that end, the EINTR from `accept()` needs to be handled gracefully,
and a sentinel variable was plumbed into the command handler.

Of note, this approach does not force "client" connections to be
immediately closed and will wait for all in-progress connections to
close before the process shuts down. Each client will close itself after
executing their next command. If a signal is repeated while the process
is waiting for one or more clients to finish, the existing behavior is
exhibited and the process terminates immediately.